// generated/openapi/models/Token.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 1inch API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Token
 * @export
 * @interface Token
 */
export interface Token {
    /**
     * 
     * @type {string}
     * @memberof Token
     */
    symbol: string;
    /**
     * 
     * @type {string}
     * @memberof Token
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Token
     */
    address: string;
    /**
     * 
     * @type {number}
     * @memberof Token
     */
    decimals: number;
    /**
     * 
     * @type {string}
     * @memberof Token
     */
    logoURI: string;
}

export function TokenFromJSON(json: any): Token {
    return TokenFromJSONTyped(json, false);
}

export function TokenFromJSONTyped(json: any, ignoreDiscriminator: boolean): Token {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'symbol': json['symbol'],
        'name': json['name'],
        'address': json['address'],
        'decimals': json['decimals'],
        'logoURI': json['logoURI'],
    };
}

export function TokenToJSON(value?: Token | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'symbol': value.symbol,
        'name': value.name,
        'address': value.address,
        'decimals': value.decimals,
        'logoURI': value.logoURI,
    };
}